+++ CHANGES.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix not displaying version in the settings screen and title on desktop
 * Fix potential memory leak when updating previews
 * Fix tutorial screen showing vertical scroll bar on high w:h ratio devices
+* Fix possibility of the async thread not being disposed properly
 
 ### Removed